+++ packages/kit/src/core/adapt/builder.spec.js
@@ -5,6 +5,7 @@ import { test } from 'uvu';
 import * as assert from 'uvu/assert';
 import glob from 'tiny-glob/sync.js';
 import { create_builder } from './builder.js';
+import { posixify } from '../../utils/filesystem.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = join(__filename, '..');
@@ -45,7 +46,11 @@ test('copy files', () => {
 	const dest = join(__dirname, 'output');
 
 	rmSync(dest, { recursive: true, force: true });
-	builder.writeClient(dest);
+
+	assert.equal(
+		builder.writeClient(dest),
+		glob('**', { cwd: dest, dot: true, filesOnly: true }).map(posixify)
+	);
 
 	assert.equal(
 		glob('**', { cwd: `${outDir}/output/client`, dot: true }),
@@ -53,7 +58,11 @@ test('copy files', () => {
 	);
 
 	rmSync(dest, { recursive: true, force: true });
-	builder.writeServer(dest);
+
+	assert.equal(
+		builder.writeServer(dest),
+		glob('**', { cwd: dest, dot: true, filesOnly: true }).map(posixify)
+	);
 
 	assert.equal(
 		glob('**', { cwd: `${outDir}/output/server`, dot: true }),