+++ packages/kit/src/utils/url.js
@@ -76,6 +76,21 @@ export function decode_params(params) {
 }
 
 /**
+ * The error when a URL is malformed is not very helpful, so we augment it with the URI
+ * @param {string} uri
+ */
+export function decode_uri(uri) {
+	try {
+		return decodeURI(uri);
+	} catch (e) {
+		if (e instanceof Error) {
+			e.message = `Failed to decode URI: ${uri}\n` + e.message;
+		}
+		throw e;
+	}
+}
+
+/**
  * URL properties that could change during the lifetime of the page,
  * which excludes things like `origin`
  * @type {Array<keyof URL>}