+++ packages/kit/test/apps/basics/test/client.test.js
@@ -544,7 +544,7 @@ test.describe('data-sveltekit attributes', () => {
 
 		const module = process.env.DEV
 			? `${baseURL}/src/routes/data-sveltekit/preload-data/target/+page.svelte`
-			: `${baseURL}/_app/immutable/components/pages/data-sveltekit/preload-data/target/_page`;
+			: `${baseURL}/_app/immutable/entry/data-sveltekit-preload-data-target-page`;
 
 		await page.goto('/data-sveltekit/preload-data');
 		await page.locator('#one').dispatchEvent('mousemove');
@@ -636,10 +636,17 @@ test.describe('Content negotiation', () => {
 	});
 });
 
-test.describe('env in app.html', () => {
-	test('can access public env', async ({ page }) => {
+test.describe('env', () => {
+	test('can access public env in app.html', async ({ page }) => {
 		await page.goto('/');
 		expect(await page.locator('body').getAttribute('class')).toContain('groovy');
+	});
+
+	test('can access public env in hooks.client.js', async ({ page }) => {
+		await page.goto('/');
+		expect(await page.evaluate(() => window.PUBLIC_DYNAMIC)).toBe(
+			'accessible anywhere/evaluated at run time'
+		);
 	});
 });