+++ packages/package/test/fixtures/exports-merge/src/lib/why_underscore.js
@@ -1,1 +1,0 @@
-// underscore in filename