+++ packages/adapter-vercel/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-vercel
 
+## 2.0.4
+
+### Patch Changes
+
+- fix: add leading / to isr routes ([#9100](https://github.com/sveltejs/kit/pull/9100))
+
+## 2.0.3
+
+### Patch Changes
+
+- fix: get ISR working on Vercel ([#9063](https://github.com/sveltejs/kit/pull/9063))
+
 ## 2.0.2
 
 ### Patch Changes