+++ packages/migrate/CHANGELOG.md
@@ -1,5 +1,17 @@
 # svelte-migrate
 
+## 1.1.1
+
+### Patch Changes
+
+- fix: include utils in migrate's published files ([#9085](https://github.com/sveltejs/kit/pull/9085))
+
+## 1.1.0
+
+### Minor Changes
+
+- feat: add `@sveltejs/package` migration (v1->v2) ([#8922](https://github.com/sveltejs/kit/pull/8922))
+
 ## 1.0.1
 
 ### Patch Changes